refactor to utils folder | ignore Element var
